# Use buffer pool in request log handler