--- a/pkg/http/request_log.go
+++ b/pkg/http/request_log.go
@@ -22,6 +22,7 @@ import (
 	"io"
 	"net/http"
 	"strings"
+	"sync"
 	"sync/atomic"
 	"text/template"
 	"time"
@@ -99,7 +100,7 @@ func NewRequestLogHandler(h http.Handler, w io.Writer, templateStr string,
 }
 
 // SetTemplate sets the template to use for formatting request logs.
-// Setting the template to an empty string turns of writing request logs.
+// Setting the template to an empty string turns off writing request logs.
 func (h *RequestLogHandler) SetTemplate(templateStr string) error {
 	var t *template.Template
 	// If templateStr is empty, we will set the template to nil
@@ -163,12 +164,20 @@ func (h *RequestLogHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	h.handler.ServeHTTP(rr, r)
 }
 
+var bufPool = sync.Pool{
+	New: func() interface{} {
+		return new(bytes.Buffer)
+	},
+}
+
 func (h *RequestLogHandler) write(t *template.Template, in *RequestLogTemplateInput) {
-	// Use a local buffer to store the whole template expansion first. If h.writer
-	// is used directly, parallel template executions may result in interleaved
-	// output.
-	w := bytes.Buffer{}
-	if err := t.Execute(&w, in); err != nil {
+	// Use a buffer to store the whole template expansion first. If h.writer is
+	// used directly, parallel template executions may result in interleaved output.
+	w := bufPool.Get().(*bytes.Buffer)
+	w.Reset()
+	defer bufPool.Put(w)
+
+	if err := t.Execute(w, in); err != nil {
 		// Template execution failed. Write an error message with some basic information about the request.
 		fmt.Fprintf(h.writer, "Invalid request log template: method: %v, response code: %v, latency: %v, url: %v\n",
 			in.Request.Method, in.Response.Code, in.Response.Latency, in.Request.URL)